refactor(design-system): combine prop, slot and emit components into one api component
