+++ packages/design-system/docs/.vitepress/composables/index.ts
@@ -1,2 +1,1 @@
 export * from './useComponentName'
-export * from './useLoadMetaData'
