+++ packages/design-system/docs/.vitepress/theme/index.ts
@@ -4,9 +4,7 @@ import * as components from './../../../src/components'
 import * as directives from './../../../src/directives'
 import './custom.scss'
 import LiveCodeBlock from '../LiveCodeBlock.vue'
-import ComponentEmits from '../components/ComponentEmits.vue'
-import ComponentProps from '../components/ComponentProps.vue'
-import ComponentSlots from '../components/ComponentSlots.vue'
+import ComponentApi from '../components/ComponentApi.vue'
 
 export default {
   extends: DefaultTheme,
@@ -15,9 +13,7 @@ export default {
     app.use(gettext)
 
     app.component('LiveCodeBlock', LiveCodeBlock)
-    app.component('ComponentEmits', ComponentEmits)
-    app.component('ComponentProps', ComponentProps)
-    app.component('ComponentSlots', ComponentSlots)
+    app.component('ComponentApi', ComponentApi)
     Object.values(components).forEach((c) => app.component(c.__name, c))
     Object.values(directives).forEach((d) => app.directive(d.name, d))
   }
