+++ packages/design-system/docs/components/OcButton/OcButton.md
@@ -89,14 +89,4 @@ A click handler can be registered via the `@click` property.
 <<< @/components/OcButton/handler.vue
 :::
 
-## Properties
-
-::: props
-
-## Emits
-
-::: emits
-
-## Slots
-
-::: slots
+::: component-api
